+++ desktop-widgets/tab-widgets/maintab.ui
@@ -491,83 +491,6 @@
        </item>
       </layout>
      </widget>
-     <widget class="QWidget" name="equipmentTab">
-      <attribute name="title">
-       <string>Equipment</string>
-      </attribute>
-      <layout class="QVBoxLayout" name="verticalLayout_10">
-       <item>
-        <widget class="QScrollArea" name="scrollArea_2">
-         <property name="frameShape">
-          <enum>QFrame::NoFrame</enum>
-         </property>
-         <property name="frameShadow">
-          <enum>QFrame::Plain</enum>
-         </property>
-         <property name="widgetResizable">
-          <bool>true</bool>
-         </property>
-         <widget class="QWidget" name="scrollAreaWidgetContents2">
-          <property name="geometry">
-           <rect>
-            <x>0</x>
-            <y>0</y>
-            <width>523</width>
-            <height>739</height>
-           </rect>
-          </property>
-          <layout class="QGridLayout" name="equipmentTabScrollAreaLayout">
-           <property name="leftMargin">
-            <number>0</number>
-           </property>
-           <property name="topMargin">
-            <number>0</number>
-           </property>
-           <property name="rightMargin">
-            <number>0</number>
-           </property>
-           <property name="bottomMargin">
-            <number>0</number>
-           </property>
-           <property name="spacing">
-            <number>2</number>
-           </property>
-           <item row="1" column="0">
-            <widget class="QWidget" name="widget" native="true">
-             <layout class="QVBoxLayout" name="cylinderWeightsLayout">
-              <property name="spacing">
-               <number>0</number>
-              </property>
-              <property name="leftMargin">
-               <number>0</number>
-              </property>
-              <property name="topMargin">
-               <number>0</number>
-              </property>
-              <property name="rightMargin">
-               <number>0</number>
-              </property>
-              <property name="bottomMargin">
-               <number>0</number>
-              </property>
-              <item>
-               <widget class="QSplitter" name="splitter">
-                <property name="orientation">
-                 <enum>Qt::Vertical</enum>
-                </property>
-                <widget class="TableView" name="cylinders" native="true"/>
-                <widget class="TableView" name="weights" native="true"/>
-               </widget>
-              </item>
-             </layout>
-            </widget>
-           </item>
-          </layout>
-         </widget>
-        </widget>
-       </item>
-      </layout>
-     </widget>
     </widget>
    </item>
   </layout>
@@ -583,12 +506,6 @@
    <class>StarWidget</class>
    <extends>QWidget</extends>
    <header>desktop-widgets/starwidget.h</header>
-   <container>1</container>
-  </customwidget>
-  <customwidget>
-   <class>TableView</class>
-   <extends>QWidget</extends>
-   <header>desktop-widgets/tableview.h</header>
    <container>1</container>
   </customwidget>
   <customwidget>
